Seamless scrolling with multiple segment matrix
Scrolling is broken at the edge of segment LED matrix module. 
Ex) if we use 2 segments LED matrix module and stroll a dot from (0,0) to right. The dot moves from (0,0) to (7,0) of 2nd segment and jumps to (0,0) of the 1st segment from (7,0) of the 2nd segment.  
The updated code makes no broken scrolling with multiple segment LED matrix